Less is more cleanup reduced 418 (354 sloc) lines to -> 299 lines 1. Removing unnecessary code decorations like: /*##################################################################### **  Included Files #####################################################################*/ because they are obvious and do not contribute to code readability/maintenance
2. Removed brief documentation
/**********************************************************************
 * @brief Main execution.
 *
 * @param [in] argc Number of arguments.
 * @param [in] argv List of arguments.
 *
 * @return Exit code.
 *
 * @note None.
**********************************************************************/

This is not a public library where the functionality and MEANING of
parameters as well as limitation on parameters must be documented.

The documentation for the sake of documentation is more noise then information.

E.g. agv[] is actually at interesting vector of pointers to the utf8 or ASCII strings
(useful knowledge which on is it depends on platform) and is modifiable (can remove
and rewrite pointers inside argv[] vector but on some platforms can point to
the strings that reside in read-only memory making strtok() call on the argument
failure prone.)

Such kind ^^^ of discussion is outside of scope of the program and sits better
in README document if we need to educate our engineer on argc/argv usage intricacies.

3. Empty lines. Code is not a poem or a song (can be).
Paragraphs are functions and it is great when they can be observed in full
in a single glance w/o excessive scrolling.
Because code has control structure with matching { and } brackets human brain
works the best when it is able to see starting and ending points of all control
structures at the same time. It is not always possible - some code is complex
but adding empty lines inside the code does not do anything for "readability".
It just makes code harder to observe.