remove unnecessary "(" ")" fixed printf() fixed my own bug introduced in previous commits.
1. C is parentheses rich language. Adding more does not make code readable.
It actually leads to more mistakes that it prevents.
Solid knowledge of operation priorities should not be substituted
by extra unnecessary parentheses.

Less is more as usual.

2. customary errors are printed to stderr.
It allows shell script to filter and redirect the errors
but even more importantly organize data pipes not pouted
with error messages passing only correct information along.

3. I am as prone to writing bug as anyone else.
            int ix = i + (stride * y);
should be
            int ix = i * stride + j;